Host backend as a Netlify Function and serve frontend + API from one site
- Extract the Express app into backend/src/app.ts so it can be shared by the
  local dev server and the serverless handler.
- Add netlify/functions/api.ts wrapping the app with serverless-http.
- Build from the repo root (npm workspaces) so backend deps install; route
  /api/* to the function ahead of the SPA fallback in netlify.toml.
