+++ backend/src/app.ts
@@ -1,0 +1,20 @@
+import express, { Request, Response, NextFunction } from 'express';
+import cors from 'cors';
+import routes from './routes';
+
+const app = express();
+
+// Middleware
+app.use(cors());
+app.use(express.json());
+
+// API routes
+app.use('/api', routes);
+
+// Error handling middleware
+app.use((err: Error, _req: Request, res: Response, _next: NextFunction) => {
+  console.error('Error:', err.message);
+  res.status(500).json({ error: 'Internal server error' });
+});
+
+export default app;
