+++ backend/src/server.ts
@@ -1,25 +1,9 @@
 import 'dotenv/config';
-import express, { Request, Response, NextFunction } from 'express';
-import cors from 'cors';
-import routes from './routes';
+import app from './app';
 
-const app = express();
 const PORT = process.env.PORT || 3000;
 
-// Middleware
-app.use(cors());
-app.use(express.json());
-
-// API routes
-app.use('/api', routes);
-
-// Error handling middleware
-app.use((err: Error, _req: Request, res: Response, _next: NextFunction) => {
-  console.error('Error:', err.message);
-  res.status(500).json({ error: 'Internal server error' });
-});
-
-// Start server
+// Start server (local development / standalone hosting)
 app.listen(PORT, () => {
   console.log(`Server running on http://localhost:${PORT}`);
 });
